+++ src/common/server.hpp
@@ -207,7 +207,7 @@ public:
 #endif
 };
 /* eventually need to keep the tcp_* functions isolated to server.c */
-int tcp_send_len (server *serv, const char *buf, int len);
+int tcp_send_len (server *serv, const char *buf, size_t len);
 void tcp_sendf (server *serv, const char *fmt, ...) G_GNUC_PRINTF (2, 3);
 int tcp_send_real (void *ssl, int sok, const char *encoding, int using_irc, const char *buf, int len, server *);
